app/functions/fetch-weather-data/requirements.txt
requests
google-cloud-firestore
functions-framework==3.*